fix: deleted the portfolio state in session, absolutly useless idk wtf i was doing
