+++ backend/app/agents/chat_agent.py
@@ -26,11 +26,10 @@ from ..models import (
     ResponseGenerationResponse,
     UIHints,
 )
+from ..models.assets import Asset, Cash, Crypto, Mortgage, RealEstate, Stock
 from .modules import (
     EntityExtractor,
-    FormPreparer,
     IntentClassifier,
-    PortfolioOperations,
     ResponseGenerator,
     WorkflowUtils,
 )
@@ -66,9 +65,7 @@ class ChatAgent:
 
         self.intent_classifier = IntentClassifier(self.llm)
         self.entity_extractor = EntityExtractor(self.llm)
-        self.portfolio_operations = PortfolioOperations()
         self.response_generator = ResponseGenerator(self.llm)
-        self.form_preparer = FormPreparer()
         self.workflow_utils = WorkflowUtils()
 
         self.session_storage = get_session_storage()
@@ -234,37 +231,15 @@ class ChatAgent:
 
     @observe(name="update_portfolio_node")
     def _update_portfolio_node(self, state: ChatAgentState) -> ChatAgentState:
-        logger.info(f"Updating in-memory portfolio based on intent: {state.intent}")
-        current_assets_count = len(state.session.portfolio_state.assets)
+        logger.info(f"Processing portfolio completion intent: {state.intent}")
+        
+        # For COMPLETE_PORTFOLIO intent, just update UI hints
+        updated_ui_hints = (state.ui_hints or UIHints()).model_copy(update={
+            "show_completion_button": True,
+            "show_portfolio_summary": True
+        })
 
-        entities_list = state.entities
-        intent = state.intent or Intent.UNCLEAR
-
-        new_ui_hints = self.portfolio_operations.update_portfolio(
-            session=state.session,
-            entities=entities_list,
-            intent=intent
-        )
-
-        updated_ui_hints = (state.ui_hints or UIHints()).model_copy()
-        if new_ui_hints.show_portfolio_summary:
-            updated_ui_hints.show_portfolio_summary = True
-        if new_ui_hints.suggest_asset_types:
-            updated_ui_hints.suggest_asset_types = True
-        if new_ui_hints.current_asset_count > 0:
-            updated_ui_hints.current_asset_count = new_ui_hints.current_asset_count
-        if new_ui_hints.show_completion_button:
-            updated_ui_hints.show_completion_button = True
-        if new_ui_hints.highlight_missing_info:
-            updated_ui_hints.highlight_missing_info = True
-
-        new_assets_count = len(state.session.portfolio_state.assets)
-
-        if new_assets_count != current_assets_count:
-            logger.info(f"In-memory portfolio updated: {current_assets_count} → {new_assets_count} assets")
-        else:
-            logger.info("Portfolio state unchanged")
-
+        logger.info("Portfolio completion UI hints updated")
         return state.model_copy(update={"ui_hints": updated_ui_hints})
 
     @observe(name="generate_response_node")
@@ -281,8 +256,7 @@ class ChatAgent:
                 session=state.session,
                 user_message=state.user_message,
                 intent=state.intent or Intent.UNCLEAR,
-                entities=state.entities,
-                portfolio_state=state.session.portfolio_state
+                entities=state.entities
             )
 
             logger.info(f"Response generated ({len(result.response)} characters)")
@@ -293,17 +267,10 @@ class ChatAgent:
 
     @observe(name="prepare_form_node")
     def _prepare_form_node(self, state: ChatAgentState) -> ChatAgentState:
-        logger.info("Preparing portfolio form for user review")
-        result = self.form_preparer.prepare_form(
-            session=state.session,
-            portfolio_state=state.session.portfolio_state
-        )
-        logger.info(f"Form prepared with {len(result.form_data.assets)} assets")
+        logger.info("Portfolio form generation no longer needed - using direct confirmation flow")
         return state.model_copy(update={
-            "show_form": result.show_form,
-            "form_data": result.form_data,
-            "response": result.response,
-            "ui_hints": result.ui_hints
+            "show_form": False,
+            "response": ResponseGenerationResponse(response="Portfolio action completed")
         })
 
     def _build_asset_confirmation(self, entities: dict, intent: Intent) -> AssetConfirmation | None:
@@ -491,11 +458,6 @@ class ChatAgent:
                 "session_id": session_id,
                 "ui_hints": dump(result.ui_hints),
                 "show_form": result.show_form,
-                "form_data": dump(result.form_data) if result.form_data else None,
-                "portfolio_summary": {
-                    "assets": len(session.portfolio_state.assets),
-                    "is_complete": session.portfolio_state.is_complete,
-                },
                 **(
                     {
                         "confirmation_request": dump(result.confirmation_request),
@@ -511,7 +473,6 @@ class ChatAgent:
                 output=response,
                 metadata={
                     "success": True,
-                    "asset_count": len(session.portfolio_state.assets),
                     "show_form": result.show_form,
                     "intent": dump(result.intent),
                     "entities_extracted": bool(result.entities),
@@ -523,8 +484,7 @@ class ChatAgent:
             )
 
             logger.info(
-                "Message processed - Assets: %s, Confirmation: %s",
-                len(session.portfolio_state.assets),
+                "Message processed - Confirmation: %s",
                 bool(result.confirmation_request),
             )
             return response
@@ -653,7 +613,7 @@ class ChatAgent:
                     portfolio_updated=False
                 )
 
-            asset = self.portfolio_operations._create_asset_from_entity(primary_entity)
+            asset = self._create_asset_from_entity(primary_entity)
 
             if not asset:
                 return PortfolioActionResult(
@@ -698,17 +658,57 @@ class ChatAgent:
                 error=str(e)
             )
 
-    def get_session_portfolio(self, session_id: str) -> Portfolio | None:
-        logger.debug(f"Retrieving portfolio for session: {session_id}")
-        session = self.session_storage.get(session_id)
-        if not session:
-            logger.warning(f"No session found for ID: {session_id}")
+    def _create_asset_from_entity(self, entity: EntityData) -> Asset | None:
+        """Create an Asset object from extracted entity data."""
+        if not entity.asset_type:
             return None
 
-        if session.portfolio_state.assets:
-            logger.info(f"Found portfolio with {len(session.portfolio_state.assets)} assets for session {session_id}")
-            return Portfolio(assets=session.portfolio_state.assets)
-        logger.info(f"No assets found in portfolio for session {session_id}")
+        asset_type = entity.asset_type.lower()
+
+        try:
+            if asset_type == "stock":
+                if entity.ticker and entity.shares:
+                    return Stock(
+                        ticker=entity.ticker.upper(),
+                        shares=float(entity.shares)
+                    )
+
+            elif asset_type in ["crypto", "cryptocurrency"]:
+                if entity.symbol and entity.amount:
+                    return Crypto(
+                        symbol=entity.symbol.upper(),
+                        amount=float(entity.amount)
+                    )
+
+            elif asset_type in ["real_estate", "realestate", "property"]:
+                if entity.address and entity.market_value:
+                    return RealEstate(
+                        address=entity.address,
+                        market_value=float(entity.market_value)
+                    )
+
+            elif asset_type == "mortgage":
+                if entity.lender and entity.balance:
+                    return Mortgage(
+                        lender=entity.lender,
+                        balance=float(entity.balance),
+                        property_address=getattr(entity, 'property_address', None)
+                    )
+
+            elif asset_type == "cash":
+                if entity.amount:
+                    return Cash(
+                        currency=entity.currency or "USD",
+                        amount=float(entity.amount)
+                    )
+
+        except (ValueError, TypeError) as e:
+            logger.error(f"Asset creation failed: {e}")
+
+        return None
+
+    def get_session_portfolio(self, session_id: str) -> Portfolio | None:
+        logger.debug(f"Portfolio state removed from sessions - no in-memory portfolio available for session: {session_id}")
         return None
 
     def clear_session(self, session_id: str):
