+++ backend/app/agents/modules/__init__.py
@@ -1,15 +1,11 @@
 from .entity_extractor import EntityExtractor
-from .form_preparer import FormPreparer
 from .intent_classifier import IntentClassifier
-from .portfolio_operations import PortfolioOperations
 from .response_generator import ResponseGenerator
 from .workflow_utils import WorkflowUtils
 
 __all__ = [
     "EntityExtractor",
-    "FormPreparer",
-    "IntentClassifier",
-    "PortfolioOperations",
+    "IntentClassifier", 
     "ResponseGenerator",
     "WorkflowUtils",
 ]
